# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stablecoin Tracker API
  version: 1.0.0
  description: |
    Interface agreement for the Stablecoin Tracker frontend.

    The contract is split in two layers, following the Single Responsibility Principle:

      1. **Master data / entities** (`/countries`, `/issuers`, `/stablecoins`,
         `/licenses`, `/reserve-types`, `/reference-assets`, `/exchange-rates`,
         `/wallets`, `/transactions`) — one resource per entity in the data model.
      2. **Analytics / aggregates** (`/analytics/*`) — read-only, pre-computed
         views that feed the dashboard (adoption heatmap, corridor flows,
         country deep dive). They never expose persistence concerns and always
         return values normalized to USD using the `USD Exchange Rates` table.

    Design principles:
      - **SRP**: each path owns one responsibility (entity CRUD vs. analytic view).
      - **OCP**: new analytics endpoints can be added without breaking existing ones.
      - **LSP / ISP**: filter parameters are componentized (`#/components/parameters/*`)
        and only the relevant ones are referenced per endpoint.
      - **DIP**: the frontend depends on these schemas, not on the underlying tables.
      - Clean code: stable resource names, consistent pagination envelope,
        ISO codes (ISO 3166-1 numeric for countries, ISO 4217 for currencies,
        ISO 8601 for dates), and a single error schema (RFC 7807-style).
  contact:
...
  license:
    name: Proprietary
servers:
...
security:
  - bearerAuth: []
paths:
...
  /countries:
    get:
      tags: [Countries]
      summary: List countries
      operationId: listCountries
      parameters:
        - $ref: '#/components/parameters/RegionQuery'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Paginated list of countries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryPage'
        default:
          $ref: '#/components/responses/Error'
  /countries/{countryId}:
    parameters:
      - $ref: '#/components/parameters/CountryIdPath'
    get:
      tags: [Countries]
      summary: Get a country by ISO 3166-1 numeric code
      operationId: getCountry
      responses:
        '200':
          description: Country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
  /stablecoins:
    get:
      tags: [Stablecoins]
      summary: List stablecoins
      operationId: listStablecoins
      parameters:
        - in: query
          name: issuerId
          schema: { type: string }
        - in: query
          name: referenceAsset
          schema: { $ref: '#/components/schemas/ReferenceAssetCode' }
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated list of stablecoins
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StablecoinPage'
  /stablecoins/{stablecoinId}:
    parameters:
      - in: path
        name: stablecoinId
        required: true
        schema: { type: string }
    get:
      tags: [Stablecoins]
      summary: Get stablecoin by id
      operationId: getStablecoin
      responses:
        '200':
          description: Stablecoin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stablecoin'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Licenses
  # ---------------------------------------------------------------------------
  /licenses:
    get:
      tags: [Licenses]
      summary: List licenses / regulatory frameworks
      operationId: listLicenses
      parameters:
        - in: query
          name: countryId
          schema: { $ref: '#/components/schemas/CountryCode' }
        - in: query
          name: type
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated list of licenses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicensePage'

  # ---------------------------------------------------------------------------
  # Reserve Types & Reference Assets
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: |
        Secret token for the private admin sync endpoints. Must match the
        `ADMIN_SYNC_TOKEN` value configured on the server (application properties).

  # ---------------------------------------------------------------------------
  # Reusable parameters
  # ---------------------------------------------------------------------------
  parameters:
    CountryIdPath:
      in: path
      name: countryId
      required: true
      schema: { $ref: '#/components/schemas/CountryCode' }
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    Sort:
      in: query
      name: sort
      description: 'Sort expression, e.g. `name,-adoption`'
      schema: { type: string }
    SearchQuery:
      in: query
      name: q
      schema: { type: string }
      description: Free-text search.
    RegionQuery:
      in: query
      name: region
      schema: { type: string }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Error:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    CountryCode:
      type: string
      description: ISO 3166-1 numeric (zero-padded up to 3 digits, e.g. "840" for the United States).
      pattern: '^[0-9]{1,3}$'
      example: '840'
    ReferenceAssetCode:
      type: string
      description: Reference asset code (ISO 4217 for fiat; symbolic for commodities/crypto).
      example: USD
...
    PageMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    Problem:
      type: object
      description: RFC 7807 problem details.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }

    # --- Entities (1:1 with the data model) --------------------------------
    Country:
      type: object
      required: [countryId, name, region]
      properties:
        countryId: { $ref: '#/components/schemas/CountryCode' }
        name: { type: string }
        region: { type: string, example: South America }
...
    Stablecoin:
      type: object
      required: [stablecoinId, name, issuerId, referenceAsset]
      properties:
        stablecoinId: { type: string }
        name: { type: string, example: USDC }
        issuerId: { type: string }
        referenceAsset: { $ref: '#/components/schemas/ReferenceAssetCode' }
    License:
      type: object
      required: [licenseId, name, type, countryId]
      properties:
        licenseId: { type: string }
        name: { type: string }
        type: { type: string, description: e.g. Money Transmitter, Trust Charter, ... }
        countryId: { $ref: '#/components/schemas/CountryCode' }
...
    CountryPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Country' }
...
    StablecoinPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Stablecoin' }
    LicensePage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/License' }
...